add policy

+++ app/Http/Controllers/API/PostController.php
@@ -42,7 +42,7 @@ class PostController extends Controller
     }
 
     public function update(Post $post, UpdatePostRequest $request){
-        // $this->authorize('update', $post);
+        $this->authorize('update', $post);
         $post->update($request->validated()); 
         $id = $post->id;
         
@@ -53,7 +53,7 @@ class PostController extends Controller
     }
 
     public function destroy(Post $post, Request $request){
-        // $this->authorize('delete', $post);
+        $this->authorize('delete', $post);
         $post->delete();
         $id = $post->id;
         Cache::forget("posts:$id");
